Skip less frames for recorded videos
It seemed the frame skip I have re-introduced by
ebf404e6afdb4616f61df0242f6913d919c225ff slightly increased
the chance of detection failures. For my 59.94 FPS video with
27 games, one game was not recognized with 3 FPS. If I changed
it to 4 FPS, all 27 games were recognized, but for more safety
this patch changes the FPS to 10. With 10 FPS the performance
improvement of the frame skip was ~70% which is slightly worse
than ~80% which was achieved by 3 FPS, but this would be a
reasonable balance between performance and precision.